+++ Source/Jobs/Controllers/JobInvoiceRateAmendmentWizardController.cs
@@ -450,7 +450,7 @@ namespace Jobs.Controllers
             vm.JobInvoiceSettings = Mapper.Map<JobInvoiceSettings, JobInvoiceSettingsViewModel>(settings);
             vm.DocTypeId = id;
             vm.DocDate = DateTime.Now;
-            vm.ProcessId = settings.ProcessId;
+            vm.ProcessId = (int)settings.ProcessId;
             var EmpId = new EmployeeService(_unitOfWork).GetEmloyeeForUser(User.Identity.GetUserId());
             if (EmpId.HasValue && EmpId.Value != 0)
                 vm.OrderById = EmpId.Value;
