+++ Source/Jobs/Controllers/JobReturnHeaderController.cs
@@ -171,7 +171,7 @@ namespace Jobs.Controllers
 
             vm.DocTypeId = id;
             vm.DocDate = DateTime.Now;
-            vm.ProcessId = settings.ProcessId;
+            vm.ProcessId = (int)settings.ProcessId;
             vm.DocNo = new DocumentTypeService(_unitOfWork).FGetNewDocNo("DocNo", ConfigurationManager.AppSettings["DataBaseSchema"] + ".JobReturnHeaders", vm.DocTypeId, vm.DocDate, vm.DivisionId, vm.SiteId);
             vm.DocumentTypeSettings = new DocumentTypeSettingsService(_unitOfWork).GetDocumentTypeSettingsForDocument(vm.DocTypeId);
             ViewBag.Mode = "Add";
